feat: js integers are 64bit thus should map to long

+++ src/main/java/jjbridge/api/value/JSType.java
@@ -12,5 +12,5 @@ import jjbridge.api.runtime.TypeResolution;
  * */
 public enum JSType
 {
-    Undefined, Null, Boolean, Integer, Double, String, External, Object, Date, Array, Function
+    Undefined, Null, Boolean, Integer, Float, String, External, Object, Date, Array, Function
 }
